resolved type errors
