+++ package.json
@@ -29,7 +29,7 @@
     "expo": "^49.0.21",
     "expo-av": "~13.4.1",
     "expo-camera": "~13.4.4",
-    "expo-document-picker": "~11.5.4",
+    "expo-document-picker": "^11.7.0",
     "expo-file-system": "~15.4.4",
     "expo-font": "~11.4.0",
     "expo-image-picker": "~14.3.2",
